WIP: Add DBUS protocol
