+++ internal/protocol/dbus.go
@@ -1,0 +1,6 @@
+package protocol
+
+
+type DBusProtocolOptions struct {}
+
+func (options *DBusProtocolOptions) NewDBusConnection() {}
