+++ internal/protocol/interface.go
@@ -1,0 +1,8 @@
+package protocol
+
+type IProtocol interface {
+	StartRoutine(string) (error)
+	DispatchMessage() (error)
+	ProcessData() (error)
+	CreateDataMessage() (error)
+}
