+++ internal/protocol/protocol.go
@@ -1,0 +1,5 @@
+package protocol
+
+func NewProtocol() (IProtocol, error){
+	return nil, nil
+}
